:bug: Remove redundant targets from defaults list and sort targets (#1354)
- Remove from "default targets" the targets already stored in the hub.
- Sort all targets by `label` so they appear alphabetically in the
select option (especially important because the whole is a concatenation
of the default and hub targets, which is why this wasn't obvious when
looking at the list from dropdown).
- Use proper `name` for defaults targets. It makes them more consistent
with struct of hub stored targets. Although those names (for the default
targets only) seems to not be used anywhere in the UI. And easier if
down the track we merge both (defaults and hub targets) in the back-end.

Resolves https://issues.redhat.com/browse/MTA-1258

Signed-off-by: Gilles Dubreuil <gdubreui@redhat.com>