:seedling: Add checks for package-lock.json to lint-staged and CI (#1365)
Following up on #1357, we now know that while a `package-lock.json` can
be valid and build upstream without a `resolved` field for each external
package, for a downstream build it is not valid. In a network detached
downstream build environment, a dependency downloader will be needed to
fetch and archive all of the project's dependencies (see cachito).
Without the lock file's resolved fields, the downloader cannot ensure
the source of the packages and will fail.

To help identify this situation, the `verify_lock.mjs` script will look
for this invalid pattern and report back.

The script will be called automatically when:

- (via lint-staged) a git commit is being created and it includes a
change to `package-lock.json`

- (via github actions) as part of CI tests when a PR is submitted or
updated

With both of these test points in place, it should be difficult to merge
a change to the lockfile that will break a downstream build.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>
Co-authored-by: Ian Bolton <ibolton@redhat.com>