:bug: Fix `useFetchAssessments` query key (#1390)
The query key for `useFetchAssessments` was set to use the same query
key as `useFetchQuestionnaires`. Setting the query keys to appropriate
values prevent a situation where the first use of one of them will cause
them both to have the same response. Having an `Assessment[]` returned
instead of a `Questionnaire[]`, or vice versa, does cause some problems.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>