:bug: Fix broken rendering of migration waves table rows, properly pass isExpandable prop in useTableControlProps (#1398)
In #1345, we changed the structure of the expandable rows in the
Migration Waves table and caused the `isExpanded` prop to be present on
all rows (which should only be on the expanded content rows), which
caused all rows to be hidden because it was false. This PR restores the
intended structure: Each row is wrapped in a `<Tbody>` with the
`isExpanded` prop, which contains two `<Tr>`s: one for the row itself
and one for the expanded content.

Also removes the `cursor: "pointer"` inline style on the `Tbody` which
was causing the whole row to appear clickable even though only the
expandable cells are clickable.

Signed-off-by: Mike Turley <mike.turley@alum.cs.umass.edu>