:bug: (Backport release-0.2) - allow svn source repos to be validated (#1346) (#1417)
- Fixes customer case 03608826 "When setting up an a new application to
be analysed by the MTA, we select Subversion as the repository type and
enter the URL with the SVN protocol. The UI tells us that this is not a
valid URL for a Subversion repository. Changing the URL protocol to HTTP
makes the error go away but our department wide Subversion server is set
to use only SVN protocol and cannot be changed. This leaves us unable to
scan an application a Subversion hosted application with a valid SVN
URL."

Adds regex to allow SVN repos to be validated when adding as a source
repo to applications

Signed-off-by: ibolton336 <ibolton@redhat.com>