Archetype detail drawer: Show archetype tags and assessment tags
Since all tags for an archetype are provided in the same field `tags`,
filter them into "archetype tags" (i.e. an empty source) and "assessment
tags" (i.e. have a non-empty source) to display them separately.

Note: The "archetype tags" are the tags manually attached to the
archetype in the new/edit form.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>