:bug: Application drawer > reports: show analysis details for successful task (#1437)
Clicking the 'View analysis details' link in the application analysis
detail drawer, reports tab should show the most current application's
task details. If no task is available for an application, "Not
available" is shown.

The link for a successful task was broken. It has been updated to use
the task id instead of the application id.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>
Co-authored-by: Ian Bolton <ibolton@redhat.com>