:sparkles: Enhance tag handling on forms (#1519)
## Summary

Enhance tag handling on forms to:
  - Display the select list options as "Tag Category Name / Tag Name"
  - Sort the options by the display name
  - Add a tooltip to the selected tag labels to show the tag's category
  - Sort the selected tag labels by the display name

Resolves: https://issues.redhat.com/browse/MTA-1533

## Details

### Enhance Tag handling on Application form, Tag and Stakeholder
handling on Archetype form, and Stakeholder handling on Assessment
Wizard

Tag fields will now:
  - Show the tag category and tag name together in the select list
  - Show the tag category and tag name on the selected label list
  - Provide a tooltip on the tag with the name of the tag's tag category
- use a complex object as the form value type and default values:
`TagItemType`
  - make sure only `Ref` values get put in the assessment payload
- use a `TagItemType[]` provided by a new query for the list of options
to work with

Stakeholders and Stakeholder Groups fields will now:
  - use `Ref` as the form value type and default values
  - make sure only `Ref` values get put in the assessment payload
- transform full data objects into items specifically formatted for use
by `Autocomplete`

This uses refactoring in `Autocomplete` to support keeping a complex id
and name object in the form values.

Form value types, default value calculations, and `onValidSubmit`
payload building have been updated as needed. Form values are using
complex objects with at least `{ id, name }` data. Default values can
push any objects that match the minimum form value complex object type.
The payload builds make sure that the form values are converted to
exactly a `Ref` object to satisfy hub's strict data marshaling.

### Refactor `Autocomplete` and `HookFormAutocomplete`

- Use a complex object for its `options` and `selections` props. This
allows for a richer experience with the typeahead component, select
list, and selected label list.

- Minimize the states in use and to replace `useEffect` in favor of
`useMemo`.

- An `AutocompleteOptionProp` allows rendering a name, a separate name
in the selected label list, and a tooltip for the selected label.

- Renamed `ItemsSelect` to `HookFormAutocomplete` since all it needs to
do now is wrap a react-hook-form around an Autocomplete

### Update application and archetype rest api functions and queries

- Update api functions `createApplication` and `createArchetype` to take
as input a `New<>` payload and return the newly created entity as a
result.

- Update query `useCreateApplicationMutation` to pass the proper values
to the `onSuccess` and `onError` callbacks.

- Update various fetch queries to return `isSuccess` so component code
can ensure at least 1 successful fetch has happened before rendering
data that depends on the fetch results.

- Add query `useFetchTagsWithTagItems` as a convenience to get a full
list of tags with tag category refs, and a `TagItemType` transform of
the tags to be useful for display.

## Screen Shots
**Manual Tags select list on Application Create:**

![image](https://github.com/konveyor/tackle2-ui/assets/3985964/79f4eeeb-120b-41c3-9a03-aa9afc08dfe3)

**Manual Tags showing existing selected tags:**

![image](https://github.com/konveyor/tackle2-ui/assets/3985964/c2fddcf4-6bf7-4d6b-9acb-6dae16c60a1d)

**Update Archetype form with existing selections for Criteria tags,
Archetype Tags, Stakeholder(s) and Stakeholder Group(s) and a mouse
hover on a selected criteria tag showing the tooltip:**

![image](https://github.com/konveyor/tackle2-ui/assets/3985964/778901cf-f7c6-4bb2-b52e-15695db8627b)

---------

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>