:sparkles: Add Tag Categories to Tag filters
On the following tables, the Tag filter has been enhanced to group
the Tag select by Tag Category:

  - Applications
  - Issues / All Issues
  - Dependencies

Tables that use Tag filters now send a map of group name (category)
to an array of `FilterSelectOptionProps` (tags).  The filter render
component `MultiselectFilterControl` has been updated to allow for
grouped and ungrouped arrays of `FilterSelectOptionProps` objects.

Existing item filtering and filter values are still operating as a
`string` or `string[]`.  The net effect means that if a tag of the same
name exists in more than one category, and one of them is selected, then
all of them will be selected.

Additional change details:
  - Type `OptionPropsWithKey` renamed `FilterSelectOptionProps` for
    more consistent type/component naming
  - Added some jsdoc to `FilterToolbar`

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>