:sparkles: Add Tag Categories to Tag filters (#1535)
Resolves: https://issues.redhat.com/browse/MTA-1534

On the following tables, the Tag filter has been enhanced to show the
Tag with Tag Category:
  - Applications
  - Issues / All Issues
  - Dependencies
  - Dependency Details Drawer / Application Table

Tables that use Tag filters now use a set of options in the format "Tag
Category Name / Tag Name". The filter render component
`MultiselectFilterControl` has been updated to have special handling of
selected values where the selection chips display the Tag Name and the
Tag Category Name is displayed as a tooltip.

Existing item filtering and filter values are still operating as a
`string` or `string[]`, but use a string that can uniquely identify each
tag. The net effect means that if a tag of the same name exists in more
than one category, they are unique. Selecting one of those Tags will
only select the tag in that specific category.

Additional change details:
  - Type `OptionPropsWithKey` renamed `FilterSelectOptionProps` for more
    consistent type/component naming
  - Added some jsdoc to `FilterToolbar`

---------

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>