Render dependency versions as links
Resolves: #1340

On the Dependencies page, detail drawer, applications table: render
the version text as a link to the maven central repository.  The link
uses the dependency's sha as the key in the maven central search.

Add component `ExternalLink` to standardize rendering links outside
of the app opening in a new tab.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>