:bug: Render dependency versions as links (#1540)
Resolves: #1340

On the Dependencies page, detail drawer, applications table: render the
version text as a link to the maven central repository. The link uses
the dependency's sha as the key in the maven central search.

Add component `ExternalLink` to standardize rendering links outside of
the app opening in a new tab.

Note: The resolution to #1338 will require a change to the application
table's query filter and response object. This will be done in PR #1536.
See https://github.com/konveyor/tackle2-hub/issues/557 for the response
object change.

Screenshot:

![image](https://github.com/konveyor/tackle2-ui/assets/3985964/50f0c144-b4fe-4ba2-a625-3c8e6ae3be6c)

---------

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>
Co-authored-by: Ian Bolton <ibolton@redhat.com>