:bug: Refactor dependencies, select row by name and provider
Resolves: #1509

The Dependencies table was setup to use the `AnalysisDependency`
`.name` as the selected row key.  Since the `.name` field is not
unique across all rows multiple rows would be erroneously selected
at the same time.  To solve the problem, a UI only unique id is
generated after fetching and used for the table selection itemId.

Change details:
  - Dependencies fetch query updated to inject a UI uid and
    return `WithUiId<AnalysisDependency>[]`.  This transform is done
    outside the react-query `useQuery()` handling.

  - Updated the table to use the `WithUiId<>` key `_ui_unique_id`
    for selection.

  - Since the button rendered in the "Found in" column had no function,
    the button was removed.

  - "Found in" column moved to the last column to match the placement
    of matching columns on the Archetypes and Issues tables.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>