:bug: Handle confirm deletes with multiple spaces (#1572)
Resolves: https://issues.redhat.com/browse/MTA-1725

When confirming the deletion of an entity by requiring the user to enter
the name of the entity, the string check needs to ignore differences in
white space.

The entity name may contain multiple spaces between words, but not show
the spaces when viewing the name outside of an input box. HTML normally
collapses multiple spaces into a single space. Also if the confirming
name gets an extra space typed, that can cause a hard to see string
mismatch. Trimming and then collapsing multiple spaces into a single
space before comparing the string avoids this problem.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>