:bug: Rename Rules tags to rules labels for better clarity (#1638)
This is to address an issue where a user expected the technology tags on
the output report to be filtered, when the filtering is based on the
labels on the rule itself. This PR just makes sure the UI and analyzer
use the same terminology

tried to keep the change minimal, can change the internal variables/div
names as well if we want.

wasn't sure if :bug: or :ghost: was more appropriate

Signed-off-by: Fabian von Feilitzsch <fabian@fabianism.us>