:bug: Broken search string when navigating from deps page (#1662)
Resolves [[Dependencies]Application inventory page shows no applications
once Accessed from
dependencies](https://issues.redhat.com/browse/MTA-2007)
Resolves https://issues.redhat.com/browse/MTA-2056

- Removes the auto applied search string when direct navigating using
the side nav. This was applying filters that were not registered within
the application table ( & other tables). I.E when navigating from the
dependencies page, the search string was including an
"applications.name" filter which is just referenced as "name" within the
applications table. Future iterations may allow us to rewrite the search
string when navigating between pages or standardizing on a specific key
for app names on the back end.
- Adds the ability to parse the url string for initial search values in
the dependencies toolbar.
- Adds associated clear filters handler to clear the url when the
toolbar clear button is pressed. This is necessary to overwrite the
initial filter values. Implemented this in dependencies page &
applications table. There may be a need to incorporate this into filter
toolbar at some point.

Signed-off-by: ibolton336 <ibolton@redhat.com>
Co-authored-by: Scott Dickerson <sdickers@redhat.com>