:sparkles: Add app name filter to archetypes page & drawer link (#1673)
There was a suggestion to update the drawer chips to a link since as the
list of chips grows, it will not be sustainable to show all chips in
drawer. The change here moves to a pre-filter for the archetypes
category in the applications table which is linkable from the archetype
drawer.
https://issues.redhat.com/browse/MTA-2283

<img width="1677" alt="Screenshot 2024-02-21 at 4 51 48 PM"
src="https://github.com/konveyor/tackle2-ui/assets/11218376/f787e32b-df98-48de-92a8-91521059df69">

---------

Signed-off-by: ibolton336 <ibolton@redhat.com>
Signed-off-by: Ian Bolton <ibolton@redhat.com>