:seedling: Swap to @rollup/wasm-node for multiarch builds (#1739)
On merge to main, #1664 showed that the normal rollup package does not
support the linux-ppc64 platform. This was not caught earlier as
building on github actions always worked with the standard `rollup`
package. The error message displayed on the multiarch build attempt is:

> Error: Your current platform "linux" and architecture "ppc64"
combination is not yet supported by the native Rollup build. Please use
the WASM build "@rollup/wasm-node" instead.

Swapping the package `rollup` to `@rollup/wasm-node` should enable the
ppc64 build.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>