:seedling: Downgrade to rollup v3 for multiarch builds (#1742)
Following up on PRs #1664 and #1739, downgrade to `rollup@^3.0`. The v4
of rollup has "native" code that runs the build, and that code does not
currently work with s390 builds [1]. Since konveyor is built for x86,
arm, ppc, and s390, all of those need to work.

When the s390 builds can work with the wasm bits, we can upgrade to
`rollup@^4.0` again.

[1] - https://github.com/rollup/rollup/issues/5354

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>