:seedling: Setup CI workflow to use same container as Dockerfile (#1804)
Lookup the "FROM .* as builder" image from the project's Dockerfile, and
use that image to run the CI unit-test job.

This will ensure the unit-test job is using the same container image
that the image builds will be using. A problem with the container should
show up quickly in the unit-testing.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>
Signed-off-by: Cherry Picker <noreply@github.com>