:seedling: Add nightly runs to the ci-repo.yml workflow (#1899)
To better support ongoing green builds, add a nightly run on main from
the repo level CI. The nightly run is embedded into the existing
`ci-repo.yml` so the PR checks, merge/push to main, and nightly runs all
use the same workflow.

This should also allow the CI repo's main README.md to have a full set
of CI badges.

Note: The codecov actions will only be run when the workflow is run for
a PR.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>