:sparkles: Test image build when build related files change
If the workflow is run from a PR, and the PR includes
a change to the `Dockerfile` or `package-lock.json`, then
run image builds for all of our target platforms.

The images are built but not pushed to any repository.

We want to be reasonably sure that any major build file changes
will not cause the image-build-and-push on PR merge workflow
to break.  Doing the image build here should show up most
problems much earlier.  For example, a npm version update
in the build container could break github action `nofiles`
or network access capabilities for the npm install.

See #1742, #1746, and #1781 for some other examples of
when this check could have caught issues before a PR
merge.

Supports: #1883

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>