:ghost: Update check for running analysis / cancel analysis  (#1885)
This PR narrows down the check for analysis tasks that are cancellable.
Also the logic is updated and flipped for "disabling" the creation of a
new analysis task. Analysis tasks should be creatable when there are no
states for tasks outside of "succeeded" or "failed". This will future
proof against any other additional task states that may be added in the
future. The "scheduled" state was just added for 0.4 & is the motivation
behind this PR ( as requested by @jortel )

Signed-off-by: Ian Bolton <ibolton@redhat.com>