:sparkles: Add application tasks status icon and popover to application table (#1985)
## Summary
On the application table's name column, add a status icon. The icon
summarizes the status of the various kinds of tasks attached to an
application (a single application has 0 or more tasks while a task acts
on a single application).

The icon will be driven off looking at the most recent task (by
`createTime`) per `kind` for an application:
  - If there are no tasks, show the application in No Tasks state
  - Else if any are "running", show the application in Running
  - Else if any are "queued", show the application in Queued state
  - Else if any are "failed", show the application as Failed
  - Else if any are "canceled", show the application as Canceled
  - Else if any are "succeeded", show the application as Success

A popover on hover will render on the application name column to display
summary information about the application's
tasks. The popover will have a table that shows the most recent tasks
for each kind of task associated to the application.
  - Task id
  - Task status icon with the task kind
  - How long ago the task was started/created (exact time in a tooltip)

Based on refactoring PR #1988 and PR #1989
Closes: #1934
Fixes: #1772

## Change Details
- Created `useDecoratedApplications()` hook to run all calculations and
cross-referencing necessary to render most of the table.

- Created `ColumnApplicationName` to handle the application name column
with the status icon and popover.

- Refactored `useFetchTasks()` hook to only do a simple sort. The
filtering based on kind is no longer necessary.

- The base `ApplicationsTable` received some basic refactoring to
utilize the content of `DecoratedApplication`

## Screen Shots
Base view with various application task statuses:
![screenshot-localhost_9000-2024 06
28-02_38_20](https://github.com/konveyor/tackle2-ui/assets/3985964/c712fc21-72f0-4529-a5bd-b0e2678910f7)

Popover details of an application with canceled tasks:
![screenshot-localhost_9000-2024 06
28-02_38_39](https://github.com/konveyor/tackle2-ui/assets/3985964/4f46159a-81a0-421a-b28a-0067f703f402)

Popover details of an application with successful tasks:
![screenshot-localhost_9000-2024 06
28-02_39_09](https://github.com/konveyor/tackle2-ui/assets/3985964/5d4bc6c3-1449-421e-b938-3bea2cf87377)

Popover details of an application with failed tasks:
![screenshot-localhost_9000-2024 06
28-02_39_43](https://github.com/konveyor/tackle2-ui/assets/3985964/421fd3e6-78fa-4689-b824-20f0ea95916a)

---------

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>